+++ textxjinja/__init__.py
@@ -5,7 +5,7 @@ import logging
 from jinja2 import Environment, FileSystemLoader
 
 
-__version__ = "0.4.0.dev0"
+__version__ = "0.4.0"
 
 logger = logging.getLogger(__name__)