+++ CHANGELOG.md
@@ -16,6 +16,7 @@ please take a look at related PRs and issues and see if the change affects you.
 
 ### Changed
 - Removed click dependency. All output goes through logging.
+- Migrated to pyproject.toml
 
 ## [0.3.0] (Released: 2022-01-13)