+++ CHANGELOG.md
@@ -14,9 +14,13 @@ please take a look at related PRs and issues and see if the change affects you.
 
 ## [Unreleased]
 
+
+## [0.4.0] (Released: 2025-05-08)
+
 ### Changed
 - Removed click dependency. All output goes through logging.
 - Migrated to pyproject.toml
+
 
 ## [0.3.0] (Released: 2022-01-13)
 
@@ -68,7 +72,8 @@ please take a look at related PRs and issues and see if the change affects you.
 - Initial release
 
 
-[Unreleased]: https://github.com/textX/textX-jinja/compare/0.3.0...HEAD
+[Unreleased]: https://github.com/textX/textX-jinja/compare/0.4.0...HEAD
+[0.4.0]: https://github.com/textX/textX-jinja/compare/0.3.0...0.4.0
 [0.3.0]: https://github.com/textX/textX-jinja/compare/0.2.1...0.3.0
 [0.2.1]: https://github.com/textX/textX-jinja/compare/0.2.0...0.2.1
 [0.2.0]: https://github.com/textX/textX-jinja/compare/0.1.0...0.2.0